refactor: move paint and shape-path builders from shape to _shape
